refactor: move ui showcase file to component layout directory
