+++ src/components/layout/index.tsx
@@ -1,482 +1,1 @@
-'use client';
-
-import * as React from 'react';
-
-import {
-  AlertDialogDescription,
-  AlertDialogContent,
-  AlertDialogTrigger,
-  AlertDialogCancel,
-  AlertDialogFooter,
-  AlertDialogAction,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialog,
-} from '@/components/ui/alert-dialog';
-import {
-  CardDescription,
-  CardContent,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-  Card,
-} from '@/components/ui/card';
-import {
-  TooltipProvider,
-  TooltipContent,
-  TooltipTrigger,
-  Tooltip,
-} from '@/components/ui/tooltip';
-import {
-  DialogDescription,
-  DialogContent,
-  DialogTrigger,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  Dialog,
-} from '@/components/ui/dialog';
-import {
-  PopoverTrigger,
-  PopoverContent,
-  Popover,
-} from '@/components/ui/popover';
-import {
-  SelectContent,
-  SelectTrigger,
-  SelectGroup,
-  SelectLabel,
-  SelectValue,
-  SelectItem,
-  Select,
-} from '@/components/ui/select';
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
-import { Terminal, Settings2, Plus, BellRing, Check } from 'lucide-react';
-import { Separator } from '@/components/ui/separator';
-import { useToast } from '@/components/ui/use-toast';
-import { ToastAction } from '@/components/ui/toast';
-import { Calendar } from '@/components/ui/calendar';
-import { Checkbox } from '@/components/ui/checkbox';
-import { Skeleton } from '@/components/ui/skeleton';
-import { Button } from '@/components/ui/button';
-import { Switch } from '@/components/ui/switch';
-import { Badge } from '@/components/ui/badge';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { cn } from '@/lib/utils';
-import { Heading } from '../ui/heading';
-
-const notifications = [
-  {
-    title: 'Your call has been confirmed.',
-    description: '1 hour ago',
-  },
-  {
-    title: 'You have a new message!',
-    description: '1 hour ago',
-  },
-  {
-    title: 'Your subscription is expiring soon!',
-    description: '2 hours ago',
-  },
-];
-
-type CardProps = React.ComponentProps<typeof Card>;
-
-const ButtonDemo = () => <Button>Button</Button>;
-
-const BadgeDemo = () => <Badge>Badge</Badge>;
-
-const AvatarDemo = () => (
-  <Avatar>
-    <AvatarImage src="https://github.com/shadcn.png" alt="@shadcn" />
-    <AvatarFallback>CN</AvatarFallback>
-  </Avatar>
-);
-
-const AlertDemo = () => (
-  <Alert>
-    <Terminal className="h-4 w-4" />
-    <AlertTitle>Heads up!</AlertTitle>
-    <AlertDescription>
-      You can add components to your app using the cli.
-    </AlertDescription>
-  </Alert>
-);
-
-const AlertDialogDemo = () => (
-  <AlertDialog>
-    <AlertDialogTrigger asChild>
-      <Button variant="outline">Show Dialog</Button>
-    </AlertDialogTrigger>
-    <AlertDialogContent>
-      <AlertDialogHeader>
-        <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-        <AlertDialogDescription>
-          This action cannot be undone. This will permanently delete your
-          account and remove your data from our servers.
-        </AlertDialogDescription>
-      </AlertDialogHeader>
-      <AlertDialogFooter>
-        <AlertDialogCancel>Cancel</AlertDialogCancel>
-        <AlertDialogAction>Continue</AlertDialogAction>
-      </AlertDialogFooter>
-    </AlertDialogContent>
-  </AlertDialog>
-);
-
-const CalendarDemo = () => {
-  const [date, setDate] = React.useState<Date | undefined>(new Date());
-
-  return (
-    <Calendar
-      mode="single"
-      selected={date}
-      onSelect={setDate}
-      className="rounded-md border"
-    />
-  );
-};
-
-const CardDemo = ({ className, ...props }: CardProps) => {
-  return (
-    <Card className={cn('w-[380px]', className)} {...props}>
-      <CardHeader>
-        <CardTitle>Notifications</CardTitle>
-        <CardDescription>You have 3 unread messages.</CardDescription>
-      </CardHeader>
-      <CardContent className="grid gap-4">
-        <div className=" flex items-center space-x-4 rounded-md border p-4">
-          <BellRing />
-          <div className="flex-1 space-y-1">
-            <p className="text-sm font-medium leading-none">
-              Push Notifications
-            </p>
-            <p className="text-sm text-muted-foreground">
-              Send notifications to device.
-            </p>
-          </div>
-          <Switch />
-        </div>
-        <div>
-          {notifications.map((notification, index) => (
-            <div
-              key={index}
-              className="mb-4 grid grid-cols-[25px_1fr] items-start pb-4 last:mb-0 last:pb-0"
-            >
-              <span className="flex h-2 w-2 translate-y-1 rounded-full bg-sky-500" />
-              <div className="space-y-1">
-                <p className="text-sm font-medium leading-none">
-                  {notification.title}
-                </p>
-                <p className="text-sm text-muted-foreground">
-                  {notification.description}
-                </p>
-              </div>
-            </div>
-          ))}
-        </div>
-      </CardContent>
-      <CardFooter>
-        <Button className="w-full">
-          <Check className="mr-2 h-4 w-4" /> Mark all as read
-        </Button>
-      </CardFooter>
-    </Card>
-  );
-};
-
-const CheckboxDemo = () => {
-  return (
-    <div className="flex items-center space-x-2">
-      <Checkbox id="terms" />
-      <label
-        htmlFor="terms"
-        className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-      >
-        Accept terms and conditions
-      </label>
-    </div>
-  );
-};
-
-const DialogDemo = () => {
-  return (
-    <Dialog>
-      <DialogTrigger asChild>
-        <Button variant="outline">Edit Profile</Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-[425px]">
-        <DialogHeader>
-          <DialogTitle>Edit profile</DialogTitle>
-          <DialogDescription>
-            Make changes to your profile here. Click save when you&apos;re done.
-          </DialogDescription>
-        </DialogHeader>
-        <div className="grid gap-4 py-4">
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label htmlFor="name" className="text-right">
-              Name
-            </Label>
-            <Input id="name" value="Pedro Duarte" className="col-span-3" />
-          </div>
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label htmlFor="username" className="text-right">
-              Username
-            </Label>
-            <Input id="username" value="@peduarte" className="col-span-3" />
-          </div>
-        </div>
-        <DialogFooter>
-          <Button type="submit">Save changes</Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-const InputDemo = () => <Input type="email" placeholder="Email" />;
-
-const LabelDemo = () => (
-  <div>
-    <div className="flex items-center space-x-2">
-      <Checkbox id="terms" />
-      <Label htmlFor="terms">Accept terms and conditions</Label>
-    </div>
-  </div>
-);
-
-const PopoverDemo = () => (
-  <Popover>
-    <PopoverTrigger asChild>
-      <Button variant="outline" className="w-10 rounded-full p-0">
-        <Settings2 className="h-4 w-4" />
-        <span className="sr-only">Open popover</span>
-      </Button>
-    </PopoverTrigger>
-    <PopoverContent className="w-80">
-      <div className="grid gap-4">
-        <div className="space-y-2">
-          <h4 className="font-medium leading-none">Dimensions</h4>
-          <p className="text-sm text-muted-foreground">
-            Set the dimensions for the layer.
-          </p>
-        </div>
-        <div className="grid gap-2">
-          <div className="grid grid-cols-3 items-center gap-4">
-            <Label htmlFor="width">Width</Label>
-            <Input id="width" defaultValue="100%" className="col-span-2 h-8" />
-          </div>
-          <div className="grid grid-cols-3 items-center gap-4">
-            <Label htmlFor="maxWidth">Max. width</Label>
-            <Input
-              id="maxWidth"
-              defaultValue="300px"
-              className="col-span-2 h-8"
-            />
-          </div>
-          <div className="grid grid-cols-3 items-center gap-4">
-            <Label htmlFor="height">Height</Label>
-            <Input id="height" defaultValue="25px" className="col-span-2 h-8" />
-          </div>
-          <div className="grid grid-cols-3 items-center gap-4">
-            <Label htmlFor="maxHeight">Max. height</Label>
-            <Input
-              id="maxHeight"
-              defaultValue="none"
-              className="col-span-2 h-8"
-            />
-          </div>
-        </div>
-      </div>
-    </PopoverContent>
-  </Popover>
-);
-
-const SelectDemo = () => (
-  <Select>
-    <SelectTrigger className="w-[180px]">
-      <SelectValue placeholder="Select a fruit" />
-    </SelectTrigger>
-    <SelectContent>
-      <SelectGroup>
-        <SelectLabel>Fruits</SelectLabel>
-        <SelectItem value="apple">Apple</SelectItem>
-        <SelectItem value="banana">Banana</SelectItem>
-        <SelectItem value="blueberry">Blueberry</SelectItem>
-        <SelectItem value="grapes">Grapes</SelectItem>
-        <SelectItem value="pineapple">Pineapple</SelectItem>
-      </SelectGroup>
-    </SelectContent>
-  </Select>
-);
-
-const SeparatorDemo = () => (
-  <div>
-    <div className="space-y-1">
-      <h4 className="text-sm font-medium leading-none">Radix Primitives</h4>
-      <p className="text-sm text-muted-foreground">
-        An open-source UI component library.
-      </p>
-    </div>
-    <Separator className="my-4" />
-    <div className="flex h-5 items-center space-x-4 text-sm">
-      <div>Blog</div>
-      <Separator orientation="vertical" />
-      <div>Docs</div>
-      <Separator orientation="vertical" />
-      <div>Source</div>
-    </div>
-  </div>
-);
-
-const SkeletonDemo = () => (
-  <div className="flex items-center space-x-4">
-    <Skeleton className="h-12 w-12 rounded-full" />
-    <div className="space-y-2">
-      <Skeleton className="h-4 w-[250px]" />
-      <Skeleton className="h-4 w-[200px]" />
-    </div>
-  </div>
-);
-
-const SwitchDemo = () => (
-  <div className="flex items-center space-x-2">
-    <Switch id="airplane-mode" />
-    <Label htmlFor="airplane-mode">Airplane Mode</Label>
-  </div>
-);
-
-const ToastDemo = () => {
-  const { toast } = useToast();
-
-  return (
-    <Button
-      variant="outline"
-      onClick={() => {
-        toast({
-          title: 'Scheduled: Catch up ',
-          description: 'Friday, February 10, 2023 at 5:57 PM',
-          action: (
-            <ToastAction altText="Goto schedule to undo">Undo</ToastAction>
-          ),
-        });
-      }}
-    >
-      Add to calendar
-    </Button>
-  );
-};
-
-const TooltipDemo = () => (
-  <TooltipProvider>
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <Button variant="outline" className="w-10 rounded-full p-0">
-          <Plus className="h-4 w-4" />
-          <span className="sr-only">Add</span>
-        </Button>
-      </TooltipTrigger>
-      <TooltipContent>
-        <p>Add to library</p>
-      </TooltipContent>
-    </Tooltip>
-  </TooltipProvider>
-);
-
-export const Layout = () => {
-  return (
-    <section className="flex flex-col gap-y-4">
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">AlertDlalog</Heading>
-        <AlertDialogDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Alert</Heading>
-        <AlertDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Avatar</Heading>
-        <AvatarDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Badge</Heading>
-        <BadgeDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Button</Heading>
-        <ButtonDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Calendar</Heading>
-        <CalendarDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Card</Heading>
-        <CardDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Checkbox</Heading>
-        <CheckboxDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Dialog</Heading>
-        <DialogDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Input</Heading>
-        <InputDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Label</Heading>
-        <LabelDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Popover</Heading>
-        <PopoverDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Select</Heading>
-        <SelectDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Separator</Heading>
-        <SeparatorDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Skeleton</Heading>
-        <SkeletonDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Switch</Heading>
-        <SwitchDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Toast</Heading>
-        <ToastDemo />
-      </div>
-      <Separator className="my-6" />
-      <div className="flex flex-col justify-center items-center gap-y-2">
-        <Heading variant="h4">Tooltip</Heading>
-        <TooltipDemo />
-      </div>
-    </section>
-  );
-};
+export { Layout } from './ui';
